move rename_argument to typeutil
