+++ knowde/_feature/_shared/cli/click_decorators/each_args/domain.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 
 import functools
-from inspect import Parameter, signature
 from typing import (
-    Any,
     Callable,
     Concatenate,
     Generic,
@@ -14,8 +12,9 @@ from typing import (
 )
 
 import click
-from makefun import create_function
 from pydantic import BaseModel
+
+from knowde._feature._shared.typeutil.func import rename_argument
 
 P = ParamSpec("P")
 T = TypeVar("T")
@@ -68,33 +67,3 @@ def each_args(
 ) -> Callable[[Wrapped], Return]:
     """Decorate with uuid completion."""
     return EachArgsWrapper(converter=converter, arg_name=arg_name)
-
-
-def rename_argument(old: str, new: str) -> Callable[[Callable], Callable]:
-    def wrapper(f: Callable[Concatenate[..., P], Any]) -> Callable:
-        sig = signature(f)
-        plist = list(sig.parameters.values())
-
-        variadics = [
-            p
-            for p in plist
-            if p.kind in {Parameter.VAR_POSITIONAL, Parameter.VAR_KEYWORD}
-        ]
-        if len(variadics) == 0:
-            raise VariadicArgumentsUndefinedError
-
-        p_old = next(p for p in plist if p.name == old)
-        p_new = p_old.replace(name=new)
-        i = plist.index(p_old)
-        plist.pop(i)
-        plist.insert(i, p_new)
-        return create_function(
-            sig.replace(parameters=plist),
-            f,
-        )
-
-    return wrapper
-
-
-class VariadicArgumentsUndefinedError(Exception):
-    """*args, **kwargs引数が定義されていないっ関数は対象外."""
