+++ knowde/_feature/_shared/cli/click_decorators/each_args/test_domain.py
@@ -1,13 +1,11 @@
 from __future__ import annotations
 
-from inspect import signature
 from uuid import UUID, uuid4
 
 import click
-import pytest
 from click.testing import CliRunner
 
-from .domain import VariadicArgumentsUndefinedError, each_args, rename_argument
+from .domain import each_args
 
 ONE = uuid4()
 
@@ -50,31 +48,3 @@ def test_with_other_option() -> None:
     )
     assert uid in result.stdout
     assert "opt" in result.stdout
-
-
-def test_rename_argument() -> None:
-    """Valid case."""
-
-    def func(
-        a: str,
-        b: int,
-        *args,  # noqa: ARG001, ANN002
-        **kwargs,  # noqa: ARG001, ANN003
-    ) -> tuple[str, int]:
-        return a, b
-
-    func2 = rename_argument("a", "replaced")(func)
-    arg1, arg2 = "xyz", 1
-    assert "replaced" not in str(signature(func))
-    assert func(arg1, arg2) == func2(arg1, arg2)
-    assert "replaced" in str(signature(func2))
-
-
-def test_invalid_rename_argument() -> None:
-    """Invalid case."""
-    with pytest.raises(VariadicArgumentsUndefinedError):
-
-        @rename_argument("a", "replaced")
-        def func(a: str, b: int) -> tuple[str, int]:
-            # def func(a: str, b: int) -> tuple[str, int]:
-            return a, b
